refactor 🔥 Remove unused pricing and contact components
- Removed `FormContact.tsx` and `page.tsx` under `/pricing/[slug]/` as they are no longer needed.
- Deleted `ChatBotPricingSection.tsx` and `MarketingPricingSection.tsx` from the landing components.
- Updated `pricing/page.tsx` to import `FormContact` from the correct path.
- Adjusted sidebar links for better clarity (`Ecommerce` -> `Templates`, `Precios` -> `Planes`).
- Modified the `TypeWritter.tsx` animation text for better readability.
