+++ src/components/ui/type-writter/TypeWritter.tsx
@@ -6,13 +6,13 @@ export const TypeWritter = () => {
     <TypeAnimation
       sequence={[
         // Same substring at the start will only be typed out once, initially
-        'Ecommerce...',
-        2000, // wait 2s before replacing "Ecommerce" with "Marketing Digital"
-        'Marketing Digital...',
+        'Ecommerces ',
+        1500, // wait 2s before replacing "Ecommerce" with "Marketing Digital"
+        'Marketing Digital',
         1500,
-        'Chatbots...',
+        'Chatbots ',
         1500,
-        'Gestión de Proyectos...',
+        'Gestión de Proyectos ',
         1000
       ]}
       wrapper="span"
